requirements.txt
# Python 3.11+
python-telegram-bot[job-queue]>=21.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
